Add default feedback, update unit tests
