+++ libs/packages/layouts/src/lib/feature/footer/footer.component.spec.ts
@@ -3,13 +3,11 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { SdsFooterComponent } from './footer.component';
 import { RouterTestingModule } from '@angular/router/testing';
 import { NavigationMode } from '@gsa-sam/components';
-import { SdsCollapseModule } from '@gsa-sam/components';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { SdsAccordionModule } from '@gsa-sam/sam-material-extensions';
-import { By } from '@angular/platform-browser';
-import { ReactiveFormsModule } from '@angular/forms';
 import { FooterLogo } from './model/FooterModel';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { SdsFeedbackModule } from '../sds-feedback/sds-feedback.module';
 
 describe('SdsFooterComponent', () => {
   let component: SdsFooterComponent;
@@ -20,10 +18,10 @@ describe('SdsFooterComponent', () => {
       imports: [
         NoopAnimationsModule,
         RouterTestingModule, 
-        SdsCollapseModule, 
         FontAwesomeModule,
-        ReactiveFormsModule,
-        SdsAccordionModule]
+        SdsAccordionModule,
+        SdsFeedbackModule,
+      ]
     })
       .compileComponents();
   }));
@@ -48,14 +46,4 @@ describe('SdsFooterComponent', () => {
     component.linkClickEvent(navItem);
     expect(component.linkEvent.emit).toHaveBeenCalledWith(navItem);
   });
-
-  it('Should emit feedback content on submit click', () => {
-    component.feedbackModel.setValue('Test Feedback');
-    const feedbackSubmitSpy = spyOn(component.feedbackSubmit, 'emit');
-
-    const feedbackSubmitButton = fixture.debugElement.query(By.css('#feedbackSubmit'));
-    feedbackSubmitButton.triggerEventHandler('click', null);
-    fixture.detectChanges();
-    expect(feedbackSubmitSpy).toHaveBeenCalledWith('Test Feedback');
-  })
 });
